requirements.txt
Django==1.9.1
django-mysql==1.0.1
django-cors-headers==1.1.0
djangorestframework==3.3.2
djangorestframework-xml==1.3.0
drf-nested-routers==0.11.1
django-oauth-toolkit==0.10.0
coverage==4.0.3
pyflakes==1.0.0
factory-boy==2.6.0
mysqlclient==1.3.7
jedi==0.9.0
#six==1.10
#django-pyodbc==0.2.8
#Sphinx==1.3.1
#sphinx-rtd-theme==0.1.8
#sphinxcontrib-httpdomain==1.4
#mock==1.3.0
#gunicorn==19.3.0
#setproctitle==1.1.9
#wddx==0.4.0
#requests==2.7.0
#factory_boy==2.5.2
#gevent==1.0.2
#sphinx_intl==0.9.7
